refactor template engine logic to separate class
