+++ extensions/wikia/PortableInfobox/services/Helpers/PortableInfoboxRenderServiceHelper.php
@@ -158,28 +158,6 @@ class PortableInfoboxRenderServiceHelper {
 	}
 
 	/**
-	 * check if item type is supported and logs unsupported types
-	 *
-	 * @param string $type - template type
-	 * @param array $templates - array of supported templates
-	 *
-	 * @return bool
-	 */
-	public function isTypeSupportedInTemplates( $type, $templates ) {
-		$isValid = true;
-
-		if ( !isset( $templates[ $type ] ) ) {
-			WikiaLogger::instance()->info( self::LOGGER_LABEL, [
-				'type' => $type
-			] );
-
-			$isValid = false;
-		}
-
-		return $isValid;
-	}
-
-	/**
 	 * Checks if europa theme is enabled and used
 	 * @return bool
 	 */
